Added logo, shutdown Script, Ip, userName, password and browser URL to a task Added functionality to open a browser after the task execution if the field browser url is setted

+++ app/src/main/java/com/example/rafa/liquidgalaxypoiscontroller/Help.java
@@ -6,6 +6,6 @@ import android.support.v7.app.ActionBarActivity;
 public class Help extends ActionBarActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView((int) R.layout.activity_help);
+        setContentView(R.layout.activity_help);
     }
 }
